Fix MSVC compile errors and warnings

+++ src/VAC/LayersWidget.cpp
@@ -382,7 +382,7 @@ void LayersWidget::updateUiFromScene_()
 {
     // Show as many existing LayerWidgets as necessary
     int numLayers = scene()->numLayers();
-    int numLayerWidgets = layerWidgets_.size();
+    int numLayerWidgets = static_cast<int>(layerWidgets_.size());
     int newNumVisibleLayerWidgets = std::min(numLayers, numLayerWidgets);
     for (int i = numVisibleLayerWidgets_; i < newNumVisibleLayerWidgets; ++i) {
         layerWidgets_[i]->show();
@@ -425,7 +425,7 @@ void LayersWidget::updateUiFromScene_()
 // Precondition: all LayerWidgets are visible
 void LayersWidget::createNewLayerWidget_()
 {
-    impl_::LayerWidget * layerWidget = new impl_::LayerWidget(layerWidgets_.size());
+    impl_::LayerWidget * layerWidget = new impl_::LayerWidget(static_cast<int>(layerWidgets_.size()));
     ++numVisibleLayerWidgets_;
     layerWidgets_.push_back(layerWidget);
     layerListLayout_->addWidget(layerWidget);
